Remove 'nphysics' module
I'm happy enough with the experience so far that I'm committed
to ripping out my half-baked physics and using nphysics for everything.
Expect commits soon doing things like re-implementing gravity.
